% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bucket-policy.r
\name{get_bucket_policy}
\alias{get_bucket_policy}
\title{Retrieve an AWS S3 Bucket Policy}
\usage{
get_bucket_policy(bucket, s3_instance = NULL)
}
\arguments{
\item{bucket}{character string specifying the name of the S3 bucket}

\item{s3_instance}{Optional AWS S3 client instance. If NULL, a new instance is created using paws.storage::s3().}
}
\value{
A list containing the parsed bucket policy with all statements and conditions
}
\description{
Retrieves and parses the policy for a specified AWS S3 bucket. The function
handles the AWS API call and converts the JSON policy to an R list structure.
}
\note{
The function will stop with an error message if it fails to retrieve the bucket policy.
}
\examples{
\dontrun{
# Get policy for a bucket named "my-bucket"
policy <- get_bucket_policy("my-bucket")

# Use an existing S3 client
s3_client <- paws.storage::s3()
policy <- get_bucket_policy("my-bucket", s3_client)
}

}
