% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_s3-upload.r
\name{s3_multipart}
\alias{s3_multipart}
\title{Perform multipart upload to S3}
\usage{
s3_multipart(local_path, s3_key, s3, bucket, part_mb = 5)
}
\arguments{
\item{local_path}{character. Path to the local file.}

\item{s3_key}{character. Key for the object in the S3 bucket.}

\item{s3}{\link[paws.storage:s3]{paws.storage} An initialized S3 client object.}

\item{bucket}{character. Target S3 bucket name.}

\item{part_mb}{numeric. Part size in megabytes (minimum 5 MB).}
}
\value{
list. Result of the complete multipart upload request.
}
\description{
Uploads a large file to an S3 bucket using the multipart upload API. Aborts the upload on failure.
}
