% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_s3-upload.r
\name{s3_upload}
\alias{s3_upload}
\title{Upload a file to S3 with retry and multipart support}
\usage{
s3_upload(local_path, s3_key, s3, bucket, max_tries = 3)
}
\arguments{
\item{local_path}{character. Path to the local file.}

\item{s3_key}{character. Key for the object in the S3 bucket.}

\item{s3}{\link[paws.storage:s3]{paws.storage} An initialized S3 client object.}

\item{bucket}{character. Target S3 bucket name.}

\item{max_tries}{integer. Maximum number of upload attempts.}
}
\value{
list. A list of upload details and result.
}
\description{
Uploads a file to an S3 bucket, using multipart upload for large files and retrying on failure.
}
